Move strprintf traits into a separate mod
